backend/requirements.txt
asgiref==3.5.2
Django==2.2.19
djangorestframework==3.12.4
pytz==2022.4
sqlparse==0.4.3
typing_extensions==4.3.0
